% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solution.R
\name{election_result}
\alias{election_result}
\title{get election result}
\usage{
election_result(county_code)
}
\arguments{
\item{county_code}{country code}
}
\value{
result the parties list and results
}
\description{
Get election result data country vice in Sweden in 2022
}
