% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solution.R
\name{election_summary}
\alias{election_summary}
\title{Summarize the Swedish parliamentary election results in 2022 in a user-friendly interface}
\usage{
election_summary()
}
\arguments{
\item{county_code}{A character string representing the county code.}
}
\description{
Summarize the Swedish parliamentary election results in 2022 in a user-friendly interface
}
