% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solution.R
\name{election_summary}
\alias{election_summary}
\title{Summarize the Swedish parliamentary election results in 2022 in a user-friendly interface}
\usage{
election_summary()
}
\value{
result the parties list and results
}
\description{
Get election result data in sweden in 2022
}
